requirements.txt
asgiref==3.8.1
Django==4.2.14
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.2
Markdown==3.6
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
